# backend/api-docs.yaml
openapi: 3.0.0
info:
  title: YouTube Clone API
  version: 1.0.0
  description: |
    Complete API documentation for YouTube Clone application.
    This API provides endpoints for user management, video uploads, and comments.
  contact:
    name: API Support
    email: support@youtubeclone.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.youtubeclone.com
    description: Production server
...
paths:
...
  /api/v1/user/signup:
    post:
      tags:
        - Users
      summary: Register a new user
      description: Create a new user account with email verification. Requires ChannelName, email, phone, password, and logo image.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - ChannelName
                - email
                - phone
                - password
                - logoUrl
              properties:
                ChannelName:
                  type: string
                  example: "My Channel"
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                phone:
                  type: string
                  example: "+1234567890"
                password:
                  type: string
                  format: password
                  example: "SecurePassword123"
                logoUrl:
                  type: string
                  format: binary
      responses:
        '201':
          description: User created successfully. Verification email sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Signup successful. Please verify your email to activate your account."
                  userId:
                    type: string
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/user/login:
    post:
      tags:
        - Users
      summary: User login
      description: Authenticate user with email and password. Returns JWT token for authenticated requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "SecurePassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  ChannelName:
                    type: string
                  email:
                    type: string
                  phone:
                    type: string
                  logoUrl:
                    type: string
                  token:
                    type: string
                    description: JWT token for authentication
                  subscriber:
                    type: number
                  subscribedChannels:
                    type: array
                    items:
                      type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/user/forgot-password:
    post:
      tags:
        - Users
      summary: Request password reset
      description: Send OTP code to user's email for password reset. OTP is valid for 10 minutes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
      responses:
        '200':
          description: OTP sent to email successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OTP sent to your email"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/user/reset-password:
    post:
      tags:
        - Users
      summary: Reset password with OTP
      description: Reset user password using OTP code received via email. OTP must be valid and not expired.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - otp
                - newPassword
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                otp:
                  type: string
                  example: "123456"
                  description: 6-digit OTP code
                newPassword:
                  type: string
                  format: password
                  example: "NewSecurePassword123"
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset successfully"
        '400':
          description: Invalid OTP or OTP expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/user/verify-email:
    post:
      tags:
        - Users
      summary: Verify email address
      description: Verify user email using the verification code sent during signup.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - code
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                code:
                  type: string
                  example: "123456"
                  description: 6-digit verification code
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email verified successfully"
        '400':
          description: Invalid verification code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/user/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      description: Get authenticated user's profile with statistics and recent videos.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  ChannelName:
                    type: string
                  email:
                    type: string
                  phone:
                    type: string
                  logoUrl:
                    type: string
                  subscriber:
                    type: number
                  subscribedChannels:
                    type: array
                    items:
                      type: string
                  createdAt:
                    type: string
                    format: date-time
                  stats:
                    type: object
                    properties:
                      totalVideos:
                        type: number
                      totalViews:
                        type: number
                      totalLikes:
                        type: number
                  recentVideos:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                        title:
                          type: string
                        thumbnailUrl:
                          type: string
                        views:
                          type: number
                        likes:
                          type: number
                        createdAt:
                          type: string
                          format: date-time
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/user/update-profile:
    put:
      tags:
        - Users
      summary: Update user profile
      description: Update user profile information. Logo is optional.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                ChannelName:
                  type: string
                  example: "Updated Channel Name"
                phone:
                  type: string
                  example: "+1234567890"
                logoUrl:
                  type: string
                  format: binary
                  description: Optional - new logo image
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Profile updated Successfully"
                  updatedUser:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/user/subscribe:
    post:
      tags:
        - Users
      summary: Subscribe to a channel
      description: Subscribe to another user's channel. Cannot subscribe to your own channel.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ChannelId
              properties:
                ChannelId:
                  type: string
                  example: "507f1f77bcf86cd799439011"
                  description: ID of the channel to subscribe to
      responses:
        '200':
          description: Subscribed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Subscribed successfully"
                  data:
                    type: object
                    properties:
                      currentUser:
                        $ref: '#/components/schemas/User'
                      subscribedUser:
                        $ref: '#/components/schemas/User'
        '400':
          description: Cannot subscribe to yourself
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== VIDEO ROUTES ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        ChannelName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        logoUrl:
          type: string
          format: uri
        subscriber:
          type: number
          default: 0
        subscribedChannels:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...